requirements.txt
semantic-kernel>=1.0.0
azure-identity
python-dotenv
pydantic>=2.0.0
numpy
pandas
matplotlib
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
websockets>=11.0.0